heuser/requirements.txt
Django==3.0.2
pytz==2019.3
sqlparse==0.3.0
mysqlclient==1.4.6
django-js-asset-1.2.2
django-ckeditor==5.9.0
django-classy-tags==1.0.0
django-cookie-law==2.0.3
pytest-runner==5.2
six==1.15.0
